+++ src/main/java/com/example/blogkar/post/entity/Post.java
@@ -22,7 +22,7 @@ public class Post {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "post_id")
-    private long postId;
+    private Integer postId;
 
     // ===========================
     // Relationships
